Add final version of Hangman.java
Fixed unintended bug that only updated the first instance of a character in the placeholder 
(eg: word is 'moose'. User types 'o', but placeholder only updates to '_ o _ _ _' when it should be '_ o o _ _'), and updated JavaDocs.

This task is from The Complete Development Bootcamp course on Udemy (link in README).